Move IDataPluginServices and stubs to NP (#51152)
* Move IDataPluginServices interface to NP

* Move stubs to NP

* Fix eslint

* code review import change
