+++ src/legacy/core_plugins/data/public/filter/filter_bar/filter_editor/phrase_suggestor.tsx
@@ -24,7 +24,7 @@ import {
   withKibana,
   KibanaReactContextValue,
 } from '../../../../../../../plugins/kibana_react/public';
-import { IDataPluginServices } from '../../../types';
+import { IDataPluginServices } from '../../../../../../../plugins/data/public';
 
 export interface PhraseSuggestorProps {
   kibana: KibanaReactContextValue<IDataPluginServices>;
