+++ src/plugins/data/common/es_query/filters/stubs/phrases_filter.ts
@@ -17,15 +17,17 @@
  * under the License.
  */
 
-import { esFilters } from '../../../../../../../../../plugins/data/public';
+import { esFilters } from '../../..';
 
-export const existsFilter: esFilters.ExistsFilter = {
+export const phrasesFilter: esFilters.PhrasesFilter = {
   meta: {
     index: 'logstash-*',
+    type: 'phrases',
+    key: 'machine.os.raw',
+    value: 'win xp, osx',
+    params: ['win xp', 'osx'],
     negate: false,
     disabled: false,
-    type: 'exists',
-    key: 'machine.os',
     alias: null,
   },
   $state: {
