+++ src/plugins/data/public/stubs.ts
@@ -19,3 +19,4 @@
 
 export { stubIndexPattern } from './index_patterns/index_pattern.stub';
 export { stubFields } from './index_patterns/field.stub';
+export * from '../common/es_query/filters/stubs';
