+++ src/plugins/data/public/types.ts
@@ -17,11 +17,9 @@
  * under the License.
  */
 
-import {
-  AutocompletePublicPluginSetup,
-  AutocompletePublicPluginStart,
-} from './autocomplete_provider/types';
-
+import { CoreStart } from 'src/core/public';
+import { IStorageWrapper } from 'src/plugins/kibana_utils/public';
+import { AutocompletePublicPluginSetup, AutocompletePublicPluginStart } from '.';
 import { FieldFormatsSetup, FieldFormatsStart } from './field_formats_provider';
 import { ISearchSetup, ISearchStart } from './search';
 import { IGetSuggestions } from './suggestions_provider/types';
@@ -44,3 +42,13 @@ export interface DataPublicPluginStart {
 
 export * from './autocomplete_provider/types';
 export { IGetSuggestions } from './suggestions_provider/types';
+
+export interface IDataPluginServices extends Partial<CoreStart> {
+  appName: string;
+  uiSettings: CoreStart['uiSettings'];
+  savedObjects: CoreStart['savedObjects'];
+  notifications: CoreStart['notifications'];
+  http: CoreStart['http'];
+  storage: IStorageWrapper;
+  data: DataPublicPluginStart;
+}
