+++ x-pack/legacy/plugins/graph/public/components/guidance_panel/guidance_panel.tsx
@@ -18,7 +18,7 @@ import { i18n } from '@kbn/i18n';
 import classNames from 'classnames';
 import { FormattedMessage } from '@kbn/i18n/react';
 import { connect } from 'react-redux';
-import { IDataPluginServices } from 'src/legacy/core_plugins/data/public/types';
+import { IDataPluginServices } from 'src/plugins/data/public';
 import {
   GraphState,
   hasDatasourceSelector,
